requirements.txt
# Core Django
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Authentication & Security
django-allauth==0.57.0
djangorestframework-simplejwt==5.3.0
python-decouple==3.8

# File handling
Pillow==10.1.0
django-storages==1.14.2

# Forms & Admin
django-crispy-forms==2.1
crispy-bootstrap5==0.7

# Utilities
python-dotenv==1.0.0
django-phonenumber-field==7.2.0
phonenumbers==8.13.25

# Email & SMS
django-ses==3.5.0
africastalking==1.2.5

# Development
django-debug-toolbar==4.2.0

# Production
gunicorn==21.2.0
whitenoise==6.6.0

# Task Queue (for notifications)
celery==5.3.4
redis==5.0.1

# SendGrid (for email)
sendgrid==6.10.0